// MyPortal.Frontend/src/app/shared/layout/menu/menu.types.ts
import { MenuItem } from 'primeng/api';
import {Me} from '../../../core/interfaces/me';

export interface AppMenuContributor {
  supports(user: Me): boolean;
  getItems(user: Me): Promise<MenuItem[]> | MenuItem[];
}

export interface LinkItem {
  label: string;
  icon?: string;
  routerLink: string | any[];
  permissionsAny?: readonly string[];
  permissionsAll?: readonly string[];
}

export interface MenuCategory {
  label: string;
  icon?: string;
  children: LinkItem[];
}